refactor: Moved all table action to a seperate module and renamed them
