+++ crates/bld_ui/src/pages/home/pipelines/actions/copy.rs
@@ -3,21 +3,22 @@ use crate::{
         button::{Button, IconButton},
         card::Card,
         colors::Colors,
+        input::Input,
     },
     context::{AppDialog, AppDialogContent, RefreshPipelines},
 };
 use anyhow::{bail, Result};
-use bld_models::dtos::PipelineQueryParams;
+use bld_models::dtos::PipelinePathRequest;
 use leptos::{html::Dialog, leptos_dom::logging, *};
 use reqwest::Client;
 
-async fn delete(name: String) -> Result<()> {
-    let params = PipelineQueryParams { pipeline: name };
+async fn copy(pipeline: String, target: String) -> Result<()> {
+    let params = PipelinePathRequest { pipeline, target };
 
     let res = Client::builder()
         .build()?
-        .delete("http://localhost:6080/v1/remove")
-        .query(&params)
+        .post("http://localhost:6080/v1/copy")
+        .json(&params)
         .send()
         .await?;
 
@@ -29,36 +30,49 @@ async fn delete(name: String) -> Result<()> {
 }
 
 #[component]
-fn PipelineTableDeleteButtonDialog(
+fn PipelineCopyButtonDialog(
     #[prop(into)] name: Signal<String>,
     #[prop(into)] app_dialog: NodeRef<Dialog>,
     #[prop()] refresh: Option<RefreshPipelines>,
 ) -> impl IntoView {
-    let delete_action = create_action(|args: &(String, Option<RefreshPipelines>)| {
-        let (id, refresh) = args.clone();
+    let name_rw = create_rw_signal(String::new());
+    let target = create_rw_signal(String::new());
+    let delete_action = create_action(|args: &(String, String, Option<RefreshPipelines>)| {
+        let (pipeline, target, refresh) = args.clone();
         async move {
-            let _ = delete(id)
+            let _ = copy(pipeline, target)
                 .await
                 .map_err(|e| logging::console_error(&e.to_string()));
             let _ = refresh.map(|x| x.set());
         }
     });
 
+    create_effect(move |_| {
+        name_rw.set(name.get());
+    });
+
     view! {
         <Card>
-            <div class="flex flex-col px-8 py-12 w-[500px] h-[300px]">
-                <div class="grow">
-                    "Are you sure you want to delete this pipeline?"
-                    <p>
-                        {move || name.get()}
-                    </p>
+            <div class="flex flex-col px-8 py-12 gap-4 w-[500px] h-[400px]">
+                <div>
+                    "Create a new copy"
+                </div>
+                <div class="grow flex flex-col gap-4">
+                    <div>
+                        <label for="pipeline">Pipeline:</label>
+                        <Input id="pipeline" disabled=true value=name_rw/>
+                    </div>
+                    <div>
+                        <label for="target">Copy:</label>
+                        <Input id="target" value=target />
+                    </div>
                 </div>
                 <div class="flex gap-x-4">
                     <Button on:click=move |_| {
-                        delete_action.dispatch((name.get(), refresh));
+                        delete_action.dispatch((name.get(), target.get(), refresh));
                         let _ = app_dialog.get().map(|x| x.close());
                     }>
-                        "Delete"
+                        "Ok"
                     </Button>
                     <Button on:click=move |_| {
                         let _ = app_dialog.get().map(|x| x.close());
@@ -72,14 +86,14 @@ fn PipelineTableDeleteButtonDialog(
 }
 
 #[component]
-pub fn PipelineTableDeleteButton(#[prop(into)] name: Signal<String>) -> impl IntoView {
+pub fn PipelineCopyButton(#[prop(into)] name: Signal<String>) -> impl IntoView {
     let app_dialog = use_context::<AppDialog>();
     let app_dialog_content = use_context::<AppDialogContent>();
     let refresh = use_context::<RefreshPipelines>();
     view! {
         <IconButton
-            icon="iconoir-bin-full"
-            color=Colors::Red
+            icon="iconoir-copy"
+            color=Colors::Zinc
             on:click=move |_| {
                 let Some(AppDialog(dialog)) = app_dialog else {
                     logging::console_error("App dialog context not found");
@@ -92,7 +106,7 @@ pub fn PipelineTableDeleteButton(#[prop(into)] name: Signal<String>) -> impl Int
                 };
 
                 let _ = content.set(Some(view! {
-                    <PipelineTableDeleteButtonDialog name=name app_dialog=dialog refresh=refresh />
+                    <PipelineCopyButtonDialog name=move || name.get() app_dialog=dialog refresh=refresh/>
                 }));
 
                 let _ = dialog.get().map(|x| x.show_modal());
