+++ crates/bld_ui/src/pages/home/pipelines/actions/mod.rs
@@ -1,0 +1,11 @@
+mod copy;
+mod delete;
+mod edit;
+mod r#move;
+mod run;
+
+pub use copy::*;
+pub use delete::*;
+pub use edit::*;
+pub use r#move::*;
+pub use run::*;
